(refactor) Lifts out depth detector specific methods from vision component to their own mixin
